% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{check_meas}
\alias{check_meas}
\title{Check all the measurements in a certain project to see if they are available or if there are problems}
\usage{
check_meas(token, meas)
}
\arguments{
\item{token}{output straight from the function \code{\link{get_token_LM}}.}

\item{meas}{object from the function \code{\link{get_meas}}.}
}
\value{
a list of two elements, the first is a list of all the available series (a list of lists), and the second is the same output than \code{\link{get_meas}} with an attached column with the availability check. The function will also print which series are not available.
}
\description{
This function is utilized to extract all the available series in a project. In many cases there might be issues, and this function might help to identify where.
}
\details{
Things here gets more complicated
}
\examples{
available <- check_meas(meas = all_meas, token = mytoken)
available_series <- available$available_series
measurements_check <- available$meas_check
}
\author{
Lorenzo Menichetti, \email{lorenzo.menichetti@luke.fi}
}
