% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{create_access_list}
\alias{create_access_list}
\title{Create access_list. The function is used only internally.}
\usage{
create_access_list()
}
\value{
A list with the following objects:
protocol = "https", ip = "chambers.ghgportal.luke.fi", ping_addr = "/api/ping/", wa_addr = "/api/workerassignments/", meas_addr = "/api/measurements/", series_addr = "/api/series/", flux_addr = "/api/flux/", auth_addr ="/api/auth/login/"
}
\description{
The object `access_list` contains all the remote addresses needed to interact with the ghgportal API. It takes no parameters, just generates the object that will be used by all other functions in this package.
}
\author{
Jani Anttila, \email{jani.anttila@luke.fi}
}
