% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{get_all_series}
\alias{get_all_series}
\title{Get all the available series and check which ar enot available}
\usage{
get_all_series(token, meas_id)
}
\arguments{
\item{token}{output straight from the function \code{\link{get_token_LM}}.}

\item{meas}{object from the function \code{\link{get_meas}}.}
}
\value{
a list of two elements, the first is a list of all the available series (a list of lists), and the second is the output returned by the API relative to that specific series when the series is not available, otherwise "available".
}
\description{
This function is utilized to extract all the available series in a project. In many cases there might be issues, and this function might help to identify where.
}
\details{
Things here gets more complicated
}
\examples{
all_series <- get_all_series(token = my_token, meas_id = meas)
all_series$all_series[which(all_series$data_check == "available")]
measurements_check <- all_series$data_check
}
\author{
Lorenzo Menichetti, \email{lorenzo.menichetti@luke.fi}
}
