% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{get_meas}
\alias{get_meas}
\title{Get the list of measurement sets}
\usage{
get_meas(token, project_id)
}
\arguments{
\item{token}{output straight from the function \code{\link{get_token_LM}}.}

\item{project_id}{id of the project the user wants to query}
}
\value{
a data frame of 17 variables, and as many rows as the measurements events
}
\description{
This function is utilized to get the list of measurement sets in a given project
}
\details{
each row corresponds to a measurement set, with the relative metadata
}
\examples{
all_meas <- get_meas(mytoken, project_id = 1)
}
\author{
Jani Anttila, \email{jani.anttila@luke.fi}
}
