% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(token)
}
\arguments{
\item{uname}{username}

\item{upass}{password}

\item{toke}{output straight from the function \code{\link{get_token_LM}}.}
}
\value{
Description of the return value.
}
\description{
This function is utilized to get the projects assigned to a certain user
}
\details{
Any additional details about the function.
}
\examples{
get_projects(mytoken)
}
\author{
Jani Anttila, \email{jani.anttila@luke.fi}
}
