% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{point_names}
\alias{point_names}
\title{Get all the point id in a series}
\usage{
point_names(series)
}
\arguments{
\item{series}{object from the function \code{\link{get_series}}.}
}
\value{
a data frame
}
\description{
This function is utilized to extract all the points (collars) ID in a series
}
\details{
Things here gets more complicated
}
\examples{
points_info <- point_names(series)
}
\author{
Jani Anttila, \email{jani.anttila@luke.fi}
}
