% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{process_single_series}
\alias{process_single_series}
\title{Reformat one single element of a series in a more readable format}
\usage{
process_single_series(series, index)
}
\arguments{
\item{series}{object from the function \code{\link{get_series}}.}
}
\value{
a list for one single element of the series
}
\description{
This function is utilized to extract all the elements (measurements time series) in a series that correspond to a specific subsite ID
}
\details{
Things here gets more complicated
}
\examples{
process_single_series(series_list=series, 11)
}
\author{
Lorenzo Menichetti, \email{lorenzo.menichetti@luke.fi}
}
