% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{subsiteID_filter}
\alias{subsiteID_filter}
\title{Filter all the elements in a series based on a subsite id}
\usage{
subsiteID_filter(series, subsiteID_value)
}
\arguments{
\item{series}{object from the function \code{\link{get_series}}.}
}
\value{
a list of lists as the function \code{\link{get_token_LM}}, but subsetted for a single subsite ID.
}
\description{
This function is utilized to extract all the elements (measurements time series) in a series that correspond to a specific subsite ID
}
\details{
Things here gets more complicated
}
\examples{
subsiteID_filter(series, subsiteID_value="DP-MC")
}
\author{
Lorenzo Menichetti, \email{lorenzo.menichetti@luke.fi}
}
