% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{subsiteID_names}
\alias{subsiteID_names}
\title{Get all the subsite id in a series}
\usage{
subsiteID_names(series)
}
\arguments{
\item{series}{object from the function \code{\link{get_series}}.}
}
\value{
a data frame
}
\description{
This function is utilized to extract all the subsite ID in a series
}
\details{
Things here gets more complicated
}
\examples{
subsite_info <- subsiteID_names(series)
}
\author{
Lorenzo Menichetti, \email{lorenzo.menichetti@luke.fi}
}
