% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_wrappers.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(token)
}
\arguments{
\item{token}{output straight from the function \code{\link{get_token_LM}}.}
}
\value{
a data.frame (which should have 9 variables:  "id", "worker", "worker_name", "project", "project_name", "author", "author_name", "date", "active"). This might change in different versions of the database, eventually contact the package maintainer.
}
\description{
This function is utilized to get the projects assigned to the user
}
\details{
Any additional details about the function.
}
\examples{
get_projects(mytoken)
}
\author{
Jani Anttila, \email{jani.anttila@luke.fi}
}
